Generic test runner: Migrate MUI V5 TextField (11/n)
[ghstack-poisoned]
