+++ packages/component-driver-mui-v5-test/src/examples/switch/index.ts
@@ -1,0 +1,5 @@
+import { IExampleUnit, ScenePart } from '@atomic-testing/core';
+import { basicSwitchExample as basicSwitch } from './BasicSwitch.examples';
+export { basicSwitchTestSuite } from './BasicSwitch.examples';
+export const basicSwitchExample = basicSwitch;
+export const switchExamples = [basicSwitchExample] satisfies IExampleUnit<ScenePart, JSX.Element>[];
